requirements.txt
flask
google-api-python-client
python-dotenv
requests
yt-dlp
opencv-python-headless
numpy
Pillow
